refactor(kiloclaw): keep OpenClaw import helpers controller-local
