+++ packages/worker-utils/package.json
@@ -7,8 +7,7 @@
     ".": "./src/index.ts",
     "./instance-id": "./src/instance-id.ts",
     "./redact-headers": "./src/redact-headers.ts",
-    "./kiloclaw-billing-observability": "./src/kiloclaw-billing-observability.ts",
-    "./openclaw-import": "./src/openclaw-import.ts"
+    "./kiloclaw-billing-observability": "./src/kiloclaw-billing-observability.ts"
   },
   "scripts": {
     "test": "vitest run",
