+++ services/kiloclaw/controller/tsconfig.json
@@ -5,10 +5,6 @@
     "moduleResolution": "bundler",
     "strict": true,
     "skipLibCheck": true,
-    "paths": {
-      "@kilocode/worker-utils": ["../../../packages/worker-utils/src/index.ts"],
-      "@kilocode/worker-utils/*": ["../../../packages/worker-utils/src/*"]
-    },
     "types": ["node"],
     "noEmit": true
   },
